+++ app/src/main/java/com/verner/healthgateway/presentation/navigation/Screen.kt
@@ -14,10 +14,11 @@ const val UID_NAV_ARGUMENT = "uid"
  */
 enum class Screen(val route: String, val titleId: Int, val hasMenuItem: Boolean = true) {
   WelcomeScreen("welcome_screen", R.string.welcome_screen, false),
+  NutritionRecords("nutrition_record_screen", R.string.nutrition_record_screen),
+  NutritionRecordDetail("nutrition_record_detail", R.string.nutrition_record_detail, false),
   ExerciseSessions("exercise_sessions", R.string.exercise_sessions),
   ExerciseSessionDetail("exercise_session_detail", R.string.exercise_session_detail, false),
   SleepSessions("sleep_sessions", R.string.sleep_sessions),
-  SleepSessionDetail("sleep_session_detail", R.string.sleep_session_detail, false),
   WeightRecords("weight_records", R.string.weight_records),
   DifferentialChanges("differential_changes", R.string.differential_changes),
   PrivacyPolicy("privacy_policy", R.string.privacy_policy, false)