+++ app/src/main/java/com/verner/healthgateway/presentation/screen/nutritionrecord/NutritionRecordScreen.kt
@@ -1,6 +1,5 @@
-package com.verner.healthgateway.presentation.screen.weightrecords
+package com.verner.healthgateway.presentation.screen.nutritionrecord
 
-import android.content.Context
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -18,24 +17,38 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.health.connect.client.records.WeightRecord
-import androidx.health.connect.client.units.Mass
+import androidx.health.connect.client.records.NutritionRecord
 import com.verner.healthgateway.R
-import com.verner.healthgateway.presentation.component.weightrecord.WeightRecordRow
+import com.verner.healthgateway.presentation.component.nutritionrecord.NutritionRecordRow
 import java.time.ZonedDateTime
 import java.util.UUID
 
+/**
+ * Shows a list of [NutritionRecord]s from today.
+ */
+@Composable
+fun mapMealType(mealType: Int?): String {
+  return when (mealType) {
+    0 -> stringResource(R.string.unknown)
+    1 -> stringResource(R.string.meal_type_breakfast)
+    2 -> stringResource(R.string.meal_type_lunch)
+    3 -> stringResource(R.string.meal_type_dinner)
+    4 -> stringResource(R.string.meal_type_snack)
+    else -> "Unknown"
+  }
+}
 
 @Composable
-fun WeightRecordScreen(
-  context: Context,
+fun NutritionRecordScreen(
+  //context: Context,
   permissions: Set<String>,
   permissionsGranted: Boolean,
-  recordsList: List<WeightRecord>,
-  uiState: WeightRecordViewModel.UiState,
+  recordsList: List<NutritionRecord>,
+  uiState: NutritionRecordViewModel.UiState,
   onImportClick: () -> Unit = {},
   onExportCsvClick: () -> Unit = {},
-  //onExportDbClick: () -> Unit = {},
+  onExportDbClick: () -> Unit = {},
+  onDetailsClick: (String) -> Unit = {},
   onError: (Throwable?) -> Unit = {},
   onPermissionsResult: () -> Unit = {},
   onPermissionsLaunch: (Set<String>) -> Unit = {},
@@ -47,21 +60,21 @@ fun WeightRecordScreen(
 
   LaunchedEffect(uiState) {
     // If the initial data load has not taken place, attempt to load the data.
-    if (uiState is WeightRecordViewModel.UiState.Uninitialized) {
+    if (uiState is NutritionRecordViewModel.UiState.Uninitialized) {
       onPermissionsResult()
     }
 
-    // The [ExerciseSessionViewModel.UiState] provides details of whether the last action was a
+    // The [NutritionRecordViewModel.UiState] provides details of whether the last action was a
     // success or resulted in an error. Where an error occurred, for example in reading and
     // writing to Health Connect, the user is notified, and where the error is one that can be
     // recovered from, an attempt to do so is made.
-    if (uiState is WeightRecordViewModel.UiState.Error && errorId.value != uiState.uuid) {
+    if (uiState is NutritionRecordViewModel.UiState.Error && errorId.value != uiState.uuid) {
       onError(uiState.exception)
       errorId.value = uiState.uuid
     }
   }
 
-  if (uiState != WeightRecordViewModel.UiState.Uninitialized) {
+  if (uiState != NutritionRecordViewModel.UiState.Uninitialized) {
     LazyColumn(
       modifier = Modifier.fillMaxSize(),
       verticalArrangement = Arrangement.Top,
@@ -88,7 +101,7 @@ fun WeightRecordScreen(
               onImportClick()
             }
           ) {
-            Text(stringResource(id = R.string.import_records))
+            Text(stringResource(id = R.string.import_exercise_session))
           }
         }
         item {
@@ -101,7 +114,7 @@ fun WeightRecordScreen(
               onExportCsvClick()
             }
           ) {
-            Text(stringResource(id = R.string.export_csv_records))
+            Text(stringResource(id = R.string.export_csv_exercise_session))
           }
         }
 //        item {
@@ -119,10 +132,14 @@ fun WeightRecordScreen(
 //        }
       }
       items(recordsList) { record ->
-        WeightRecordRow(
-          ZonedDateTime.ofInstant(record.time, record.zoneOffset),
+        NutritionRecordRow(
+          ZonedDateTime.ofInstant(record.startTime, record.startZoneOffset),
           record.metadata.id,
-          Mass.kilograms(record.weight.inKilograms)
+          mapMealType(record.mealType),
+          record.energy,
+          onDetailsClick = { uid ->
+            onDetailsClick(uid)
+          }
         )
       }
     }